Add Anchor model and migrations.
Anchors are fundamental to the 3taps polling API. They keep track of
what we've polled so far: http://docs.3taps.com/polling_api.html.

An Anchor contains two things: a string representing the actual anchor,
and an integer representing the number of listings polled since the last
anchor.

The Anchor is manually initialized in `seeds.rb' to `1313701192'. This
was a suitable number deterined through manually curling the 3taps API.