Add rake task for pulling CL properties from 3taps.
Run `rake cl_properties:update' to pull in all new CL listings with the
proper filters. Every update pulls in 1000 listings at max.

After listings are polled, they're added one by one to ClProperty, with
ClImage, ClLocation, and ClAnnotation being filled in as well.

Properties without prices are discarded, as are properties which are
actually `wanted' ads.

If the property already exists in the db, it is destroyed, then added
again of not `post["deleted"]'.