Add location scoring features.
On the save of a ClLocation, the `score' function is called, which
searches for the nearest 3 grocery stores, as well as the nearest 10
restaurants. Then, it creates a LocationVenue for each of the results,
thereby caching them (so we don't need to query Yelp again, when we
display the ClProperty).

It then scores the locations, by category. The score is calculated by
using a rough approximation of distance * (speed per mode of transport),
along with some fixed constants added in (e.g.takes 3 minutes to get in your
car and find parking), as well as the average rating of the venues.

Then, after a bit of error checking, everything is saved.